% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{em_bm}
\alias{em_bm}
\title{Estimate the parameters of a binomial mixture model using the EM algorithm}
\usage{
em_bm(x, n, p1 = 0.5, p2 = 0.6, theta1 = 0.7, max_iter = 1000L, tol = 0.001)
}
\description{
Estimate the parameters of a binomial mixture model using the EM algorithm
}
