% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{export_dt}
\alias{export_dt}
\alias{export_df}
\alias{export_pval}
\alias{export_binary}
\alias{export_af}
\title{Export the mutation matrix}
\usage{
export_dt(
  mtmutObj,
  loc_list,
  p_threshold = 0.05,
  percent_interp = 0.2,
  n_interp = 2,
  min_cell_n = 0,
  p_adj_method = "fdr",
  af_threshold = 1,
  all_cell = F
)

export_df(mtmutObj, loc_list, ...)

export_pval(mtmutObj, loc_list, memoSort = F, ...)

export_binary(mtmutObj, loc_list, memoSort = F, ...)

export_af(mtmutObj, loc_list, memoSort = F, ...)
}
\arguments{
\item{mtmutObj}{The scMtioMut object}

\item{loc_list}{The list of loci}

\item{p_threshold}{The p value threshold}

\item{min_cell_n}{The minimum number of cells with mutation}

\item{p_adj_method}{The method to adjust p value}
}
\value{
The mutation matrix
}
\description{
Export the mutation matrix
}
\examples{
# d_plot = export_p(x, loc_list)
# d_plot_af = export_af(x, loc_list)
# d_plot_bin = export_binary(x, loc_list, 0.05)
}
