% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{format.mtmutObj}
\alias{format.mtmutObj}
\title{Mitochondiral mutation object}
\format{
A list with the following elements:
\describe{
\item{h5f}{H5 file handle}
\item{mut_table}{allele count table H5 group handle}
\item{loc_list}{list of available loci}
\item{loc_selected}{selected loci}
\item{cell_list}{list of available cell ids}
\item{cell_selected}{selected cell ids}
}
}
\usage{
\method{format}{mtmutObj}(x)
}
\description{
A S3 class for handling mitochondrial mutation data
}
