% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_pval}
\alias{get_pval}
\title{Get p-value list for single locus}
\usage{
get_pval(mtmutObj, loc, model = "bi", method = "fdr")
}
\arguments{
\item{mtmutObj}{a mtmutObj object}

\item{loc}{a string of the locus}

\item{method}{a string of the method for p-value adjustment}
}
\description{
Get p-value list for single locus
}
