% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{mut_filter}
\alias{mut_filter}
\title{Filter and visualize mutations}
\usage{
mut_filter(
  mtmutObj,
  min_cell = 1,
  p_threshold = 0.05,
  p_adj_method = "fdr",
  af_threshold = 1
)
}
\arguments{
\item{mtmutObj}{a mtmutObj object}

\item{min_cell}{a integer of the minimum number of cells with mutation}

\item{p_threshold}{a numeric of the p-value threshold}

\item{p_adj_method}{a string of the method for p-value adjustment,
refer to \code{\link[p.adjust]{p.adjust}}}

\item{af_threshold}{a numeric of the majority allele frequency threshold,
the major allele af < af_threshold will be considered as mutation.
The default is 1, which means no filtering}
}
\description{
Filter and visualize mutations
}
