% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{parse_mgatk}
\alias{parse_mgatk}
\title{Load mtGATK output}
\usage{
parse_mgatk(dir, prefix, h5_file = "mut.h5")
}
\arguments{
\item{dir}{a string of the mtGATK output \code{final} fold.}

\item{prefix}{a string of the prefix of the mtGATK output directory.}

\item{h5_file}{a string of the output h5 file directory.}
}
\value{
a string of the output h5 file directory.
}
\description{
This function loads the mtGATK output and save it to a H5 file.
}
\examples{
f = system.file("extdata", "mini_mgatk_out", package = "scMitoMut")
h5_f = load_mgatk(paste0(f, "/out/"), prefix = "sample", h5_file = "./mut.h5")
h5_f
x <- open_h5_file(h5_f)
x  
##
}
