% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_static.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap plot}
\usage{
plot_heatmap(mtmutObj, type = "p", cell_ann = NULL, ann_colors = NULL, ...)
}
\arguments{
\item{mtmutObj}{an object of class "mtmutObj".}

\item{type}{a string of plot type, "p" for p-value, "af" for allele frequency.}

\item{cell_ann}{a data.frame of cell annotation, with rownames as cell barcodes.}

\item{ann_colors}{a vector of colors for cell annotation with cell annotation as names.}

\item{...}{other parameters for \code{\link{export_df}} and \code{\link{pheatmap::pheatmap}}.}
}
\description{
Heatmap plot
}
