% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_static.R
\name{plot_locus}
\alias{plot_locus}
\title{Draw depth ~ AF scatter plot}
\usage{
plot_locus(d_select_maj_base, p, p_threshold = 0.05, loc = NA, maj_base = NA)
}
\description{
Draw depth ~ AF scatter plot
}
