% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_static.R
\name{plot_locus_profile}
\alias{plot_locus_profile}
\title{QC plot: 2D scatter plot for coverage ~ AF and UMAP}
\usage{
plot_locus_profile(
  loc,
  mtmutObj,
  seuratObj,
  p_threshold = 0.05,
  p_adj_method = "fdr"
)
}
\arguments{
\item{loc}{a string of genome location, e.g. "mt1000"}

\item{mtmutObj}{an object of class "mtmutObj"}

\item{p_threshold}{a numeric value of p-value threshold}

\item{p_adj_method}{a string of p-value adjustment method}
}
\description{
QC plot: 2D scatter plot for coverage ~ AF and UMAP
}
\examples{
##
}
