% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{format.mtmutObj}
\alias{format.mtmutObj}
\alias{print.mtmutObj}
\alias{is.mtmutObj}
\title{Print mtmutObj object}
\usage{
\method{format}{mtmutObj}(x, ...)

\method{print}{mtmutObj}(x, ...)

is.mtmutObj(x)
}
\arguments{
\item{x}{a mtmutObj object.}

\item{...}{other parameters passed to \code{\link[base]{format}} or \code{\link[base]{print}}.}
}
\value{
a string
}
\description{
The print method for mtmutObj object.
}
