% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_betabinomial.R
\name{process_locus_bb}
\alias{process_locus_bb}
\title{Fit one location}
\usage{
process_locus_bb(loc, mtmutObj, maj_base = NULL, max_try = 1)
}
\description{
Fit one location
}
\examples{

process_locus("mt1000", mtmutObj, maj_base = NULL, max_try = 1)

##
}
