% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_ensemble.R
\name{process_locus_bmbb}
\alias{process_locus_bmbb}
\title{Fit binomial mixture model for one locus}
\usage{
process_locus_bmbb(mtmutObj, loc, maj_base = NULL, ...)
}
\arguments{
\item{mtmutObj}{a mtmutObj object}

\item{loc}{string given the locus name (e.g. "mt1000")}

\item{maj_base}{string given the major base (e.g. "A"), if NULL auto detect the major base}
}
\value{
list of p-value and model parameters
}
\description{
Fit binomial mixture model for one locus
}
