% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMitoMut.R
\docType{package}
\name{scMitoMut}
\alias{scMitoMut-package}
\alias{scMitoMut}
\title{scMitoMut}
\description{
scMitoMut is a tool for detecting mitochondrial mutations in single-cell sequencing data.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/wenjie1991/scMitoMut}
  \item Report bugs at \url{https://github.com/wenjie1991/scMitoMut/issues}
}

}
\author{
\strong{Maintainer}: Wenjie Sun \email{sunwjie@gmail.com} (\href{https://orcid.org/0000-0002-3100-2346}{ORCID})

Other contributors:
\itemize{
  \item Leila Perie \email{leila.perie@curie.fr} [contributor]
}

}
