% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{export_dt}
\alias{export_dt}
\alias{export_df}
\alias{export_pval}
\alias{export_binary}
\alias{export_af}
\title{Export the mutation matrix}
\usage{
export_dt(mtmutObj, percent_interp = 1, n_interp = 3, all_cell = F)

export_df(mtmutObj, ...)

export_pval(mtmutObj, memoSort = F, ...)

export_binary(mtmutObj, memoSort = F, ...)

export_af(mtmutObj, memoSort = F, ...)
}
\arguments{
\item{mtmutObj}{The scMtioMut object.}

\item{percent_interp}{A numeric value, the overlapping percentage threshold for triggering interpolation. The default is 1, which means no interpolation.}

\item{n_interp}{A integer value, the minimum number of overlapped cells with mutation for triggering interpolation.}

\item{all_cell}{A boolean to indicate whether to include all cells or only cells with mutation.}
}
\value{
data.frame, data.table or matrix
}
\description{
The helper functions to export the mutation results for further analysis. The output format can be data.frame, data.table or matrix for p value, allele frequency or binary mutation status.
}
\examples{
h5_f = system.file("extdata", "mut.h5", package = "scMitoMut")
## open the h5 file and create a mtmutObj object
x <- open_h5_file(f_h5)
run_model_fit(x)
x = filter_loc(x , min_cell = 5, model = "bb", p_threshold = 0.05, p_adj_method = "fdr", af_threshold = 0.05)
x
export_df(x)
export_p(x)
export_af(x)
export_binary(x)
}
