% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{export_dt}
\alias{export_dt}
\alias{export_df}
\alias{export_pval}
\alias{export_binary}
\alias{export_af}
\title{Export the mutation matrix}
\usage{
export_dt(mtmutObj, percent_interp = 1, n_interp = 3, all_cell = F)

export_df(mtmutObj, ...)

export_pval(mtmutObj, memoSort = F, ...)

export_binary(mtmutObj, memoSort = F, ...)

export_af(mtmutObj, memoSort = F, ...)
}
\arguments{
\item{mtmutObj}{The scMtioMut object.}

\item{percent_interp}{A numeric value, the overlapping percentage threshold for triggering interpolation. The default is 1, which means no interpolation.}

\item{all_cell}{A boolean to indicate whether to include all cells or only cells with mutation.}
}
\value{
data.frame, data.table or matrix
}
\description{
Export the mutation matrix
}
\examples{
# d_plot = export_p(x, loc_list)
# d_plot_af = export_af(x, loc_list)
# d_plot_bin = export_binary(x, loc_list, 0.05)
}
