% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{mut_filter}
\alias{mut_filter}
\title{Filter and visualize mutations}
\usage{
mut_filter(
  mtmutObj,
  min_cell = NULL,
  model = NULL,
  p_threshold = NULL,
  p_adj_method = NULL,
  af_threshold = NULL
)
}
\arguments{
\item{mtmutObj}{a mtmutObj object}

\item{min_cell}{a integer of the minimum number of cells with mutation}

\item{model}{a string of the model for mutation calling, it can be "bb", "bm" or "bi"}

\item{p_threshold}{a numeric of the p-value threshold}

\item{p_adj_method}{a string of the method for p-value adjustment,
refer to \code{\link[p.adjust]{p.adjust}}}

\item{af_threshold}{a numeric of the majority allele frequency threshold,
the major allele af < af_threshold will be considered as mutation.
The default is 1, which means no filtering}
}
\value{
a mtmutObj object with loci selected
}
\description{
Filter and visualize mutations
}
