% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{parse_mgatk}
\alias{parse_mgatk}
\title{Load mtGATK output save to H5 file}
\usage{
parse_mgatk(dir, prefix, h5_file = "mut.h5")
}
\arguments{
\item{dir}{a string of the mtGATK output \code{final} fold}

\item{prefix}{a string of the prefix of the mtGATK output directory}

\item{h5_file}{a string of the output h5 file directory}
}
\value{
a string of the output h5 file directory
}
\description{
Load mtGATK output save to H5 file
}
\examples{

## Not run
# y = load_mgatk("./mgatk_out/final/", prefix = "prefix_name", h5_file = "./mut.h5")
# print(y)
# > [1] './mut.h5',

##
}
