% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{parse_mgatk}
\alias{parse_mgatk}
\title{Load mgatk}
\usage{
parse_mgatk(dir, prefix, h5_file = "mut.h5")
}
\description{
Load mgatk
}
\examples{

y = load_mgatk(mgatk_output_dir, prefix = "lib_id", h5_file = "./mut.h5")

## cell list
y$cell_id

## test h5 file read out
h5f = H5Fopen(y$h5_file)
h5f
h5g = h5f&"mut_table"
h5g$x1600
h5f$cell_id_list
h5closeAll()

##
}
