% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_static.R
\name{plot_af_coverage}
\alias{plot_af_coverage}
\title{QC plot: 2D scatter plot for coverage ~ AF}
\usage{
plot_af_coverage(
  mtmutObj,
  loc,
  model = NULL,
  p_threshold = NULL,
  p_adj_method = NULL
)
}
\description{
QC plot: 2D scatter plot for coverage ~ AF
}
