% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_static.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap plot}
\usage{
plot_heatmap(
  mtmutObj,
  loc_list,
  cell_ann = NULL,
  ann_colors = NULL,
  type = "p",
  ...
)
}
\arguments{
\item{mtmutObj}{an object of class "mtmutObj"}

\item{loc_list}{a vector of mt genome location}

\item{cell_ann}{a vector of cell annotation}

\item{ann_colors}{a vector of cell annotation colors}

\item{type}{a string of plot type, "p" for p-value, "af" for allele frequency}

\item{...}{other parameters for \code{\link{export_df}}}
}
\description{
Heatmap plot
}
