% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_betabinomial.R
\name{process_locus_bb}
\alias{process_locus_bb}
\title{Fit beta-binomial distribution for one locus}
\usage{
process_locus_bb(
  d_select_maj_base,
  selected_maj_cell = NULL,
  max_theta = 1e+06,
  max_iter = 100,
  tol = 0.001
)
}
\arguments{
\item{d_select_maj_base}{data.frame of one locus}

\item{selected_maj_cell}{vector of selected cell}

\item{max_theta}{maximum theta}

\item{max_iter}{maximum iteration}

\item{tol}{tolerance of log likelihood to stop iteration}
}
\value{
list of p-value and model parameters
}
\description{
Fit beta-binomial distribution for one locus
}
\examples{

process_locus("mt1000", mtmutObj, maj_base = NULL, max_try = 1)

##
}
