% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{subset_cell}
\alias{subset_cell}
\alias{subset_loc}
\title{Subset cell and loci}
\usage{
subset_cell(mtmutObj, cell_list)

subset_loc(mtmutObj, loc_list)
}
\arguments{
\item{mtmutObj}{a mtmutObj object}

\item{cell_list}{a list of cell barcodes}

\item{loc_list}{a list of loci}
}
\value{
a mtmutObj object with cell and loci selected
}
\description{
Subset cell and loci
}
